firsts steps of refactor
