+++ flutter_app/lib/widgets/change_mode_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:dart_mavlink/dialects/common.dart';
-import 'package:imacs/mavlink_communication.dart';
+import 'package:imacs/modules/mavlink_communication.dart';
 
 /// Define the MavMode constants and their string representations.
 const Map<int, String> mavModes = {
